feat: fixed class imports       - adding an Enum of days of the week       - Switching to LocalDateTime instead of LocalDate

+++ src/main/java/codereview/school_mate/model/DailyTime.java
@@ -13,12 +13,14 @@ public class DailyTime {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
+
     @Column(name = "start_time", nullable = false)
     private LocalTime startTime;
 
     @Column(name = "end_time", nullable = false)
     private LocalTime endTime;
 
+    //TODO Сделать LONG тип
     @Column(name = "lesson_number", nullable = false)
-    private Integer lessonNumber;
+    private Long lessonNumber;
 }
